Initial commit with minimal website
Summary:
This commit contains a relatively minimal website skeleton, with
server-side rendering, server-side and client-side routing, static site
generation, ES6 transpilation, and linting. It's not quite as minimal as
I'd //like//, but that's because it's difficult to convince oneself that
all the pieces fit together properly without having…well, a bunch of
code to glue them all together.

Test Plan:
Run `npm start` and visit http://localhost:8080/webpack-dev-server/.
Note that you can navigate with the ugly navbar, and that the page loads
instantly (because it's routing on the client side with the history
API). Make sure that the image loads. Then manually change the URL to
/webpack-dev-server/projects, and watch it do a full reload to the right
page. Next, make some changes and watch the correct page refresh
automatically. Then do a full reload of the home page, and click to the
projects page; make sure that the word "useful" is styled green and
underlined (this style should have been added by Aphrodite when loading
the second page).

Finally, kill your dev server, then run `npm run build`. Open
http://localhost:8000, and note that the same site is visible (and that
routing works). (As a sanity check, make sure that the relevant files
appear in `dist/` as well.)