Add resource preloading for images
Summary:
This patch adds a hook for registering images to be preloaded when the
link to a given route is hovered.

Test Plan:
Run `npm run serve`. In both Chrome and Firefox, open
http://localhost:8080 with the network panel open. Note that just
`me.png` is loaded. Hover over the "Projects" link, and watch the four
images be preloaded. Click the "Projects" link, and note that there are
no new network requests when the page is loaded, and that the images
load instantly. Then, do a full page refresh. Note that just the four
images on the projects page load. Hover the "Home" link and watch the
file `me.png` be preloaded similarly.