Implement the skills page
Summary:
This has a different layout than the rest of the pages, because there
are more skills than would fit comfortably in `Blurb`s. So instead it
has a dynamic master-detail view selector that shows all skills' icons,
and shows one skill's contents at a time.

Test Plan:
I've tested this for accessibility with `orca` in Firefox. I can tab
through all the links and activate them with Enter; `orca` automatically
switches to the newly added/updated div and reads it, without losing
focus on the button. (This is exactly the intended behavior.)