Use an Object.assign polyfill
Summary:
I was running into facebook/react#6451 even though I'm on React 15.3.1
and `npm ls | grep object-assign` shows only `object-assign@4.1.0`. A
bit of debugging shows that the polyfill wasn't being applied early
enough in the node bundle that's used for server-side rendering; we
have to be sure to include it before any React components are used, so
we might as well make it the very first thing!

Test Plan:
Load localhost:8080/skills and see that there aren't any SSR errors.