Use npm `prop-types`, not `React.PropTypes`
Summary:
This resolves a v15 deprecation warning by fixing all the occurrences in
my code. The recent upgrade of react-router to v3 fixes all the
occurrences in my dependencies.

Test Plan:
Run `npm start`. Note that there is a deprecation warning in the
console, and also one in the JS console of the browser once you load the
page. Note that both these warnings disappear after this commit is
applied.