Replace `dedent` with a template-only dedent
Summary:
The `dedent` module has two problems that require its replacement.
First, it dedents the substitutions in addition to the template, but the
substitutions need to be treated as opaque. This is because one such
substitution is the whole HTML contents, which could contain important
leading whitespace in a `<pre>` tag. Second, it replaces the character
sequence "backslash, newline, zero-or-more-whitespace" with the empty
string, and there is no way to disable this behavior, so there is
consequently no way to emit a non-final line ending in a backslash. This
means that it will not be possible to use `dedent` to dedent code
fragments.

Both of these issues are solved by implementing a template tag function
that serves the intended purpose.

Test Plan:
Add

    <test>${"\n    indented!"}</test>\

to the trailing backslash in `server.js`, and check the rendered output
to see that both issues mentioned above are fixed.