Don't hard-wrap freeform text [noblame]
Summary:
This commit modifies all freeform text in source code so that it has
unbounded line length instead of wrapping at 80 columns. This is what
I have always done for TeX in source control (and elsewhere) and it
works very well. It will have the following two positive effects here:
  - blames will be better isolated to their actual changes, as they will
    not cause reflow;
  - fewer `{" "}`s will be needed because there will be fewer tags at
    the starts or ends of physical lines.

I found a number of typos in this process, which I did not fix.