Use Unicode literals where possible [noblame]
Summary:
This commit modifies all occurrences of `&entity;` that refer to a
printable, non-whitespace Unicode character by replacing those
occurrences with the actual character. This is in accordance with Google
style.

Created by running the vimscript command

    :let subs = [['&rsquo;', '’'], ['&lsquo;', '‘'], ['&rdquo;', '”'], ['&ldquo;', '“'], ['&middot;', '·'], ['&mdash;', '—'], ['&ndash;', '–'], ['&hellip;', '…']] | for [x, y] in subs | silent! exe "norm :%s/" . x . "/" . y . "/g^M" | endfor | exe 'norm /&\(nbsp;\)\@!\(amp;\)\@!^M'

(replace ^M with a literal 0xD via <C-V><C-M>) on each modified file.
I already know all the vim digraphs for the above characters (plus most
of their Unicode code points…) so this won't be a maintenance burden in
the future. The only matches for the final searches were `&&`, which is
of course fine.