Rename "thoughts" to "posts"
I didn't want to call it a "blog" to avoid giving the impression of
structure or timely posting, but I couldn't really think of a good
name at the time. I ended up going with "thoughts" just because I wanted
to ship the thing. "Posts" is obviously a far better name, and less
pretentious-sounding, too!

The only downside of this change is that it breaks external permalinks.
An immediately subsequent commit will fix this.