Institute redirects
Summary:
The previous commit introduced a URL change. To preserve those
permalinks, we introduce redirect functionality, and add the relevant
redirections.

This unfortunately means that we have to fix a static domain name, but
this is not actually a practical issue. (It was just nice that
everything was domain-agnostic until now.)

Test Plan:
In `server.js`, change the meta-refresh timer from `0` to `60`. Disable
JavaScript in your browser. Load `/thoughts` or a `/thoughts/` post, and
verify that the redirection page appears correct. Undo the two changes
above (the timer change and JavaScript disabling) and verify that for
each one the redirect is as expected. Finally, verify that the navbar
and existing pages continue to render correctly.