Fix in-app links inside noscript contexts
Summary:
These were rendering without `href`, which made them unclickable. See
comments in this patch for explanation of the root cause and fix.

Test Plan:
Visit `/skills` in a text-based browser like `lynx` or a graphical
browser with JavaScript disabled. Note that in-app links, like those
under the “Java” section, are now clickable and resolve correctly.