Add support for KaTeX
Neither `react-latex` nor `react-katex` describes support for
server-side rendering, so I’ve written my own. We use a Context to
detect whether KaTeX is used on a page, and if so inject the stylesheet
into the static page. When a KaTeX element is dynamically rendered into
a page that didn’t start with the stylesheet, we add it at runtime. This
way, the HTML payload size is only increased for pages that actually
need it.

Test Plan:
Add some KaTeX to a post, like this:

```diff
diff --git a/src/pages/posts/data/autoincrementingIdTable.js b/src/pages/posts/data/autoincrementingIdTable.js
index 70fc654..bd23fde 100644
--- a/src/pages/posts/data/autoincrementingIdTable.js
+++ b/src/pages/posts/data/autoincrementingIdTable.js
@@ -2,11 +2,12 @@ import React from 'react';

 import dedent from '../../../dedent';
 import Prism from '../../../extern/Prism';
-import {CodeBlock} from '../../../Components';
+import {CodeBlock, Katex} from '../../../Components';

 function render(Title, Section) {
     return <article>
         <Title>A cute autoincrementing ID table in one line of Python</Title>
+        Recall that <Katex display={true} tex="\sqrt{a^2 + b^2} = c" /> and you will do okay, easy as <Katex tex="\pi" />.
         <p>
             Consider the following definition:
         </p>
```

then navigate to that page. Then, navigate to a different post,
hard-reload, and navigate back to the page with the TeX without
reloading; note that it still appears properly.