Make wide equations responsive
We now include two copies of the wider equations, one with an `aligned`
layout so that it’s much less wide.

A downside of this is that the equations appear duplicated in Lynx,
which appears to not respect the conditional `display: none` on either
form. But I’m willing to live with that.